refactor: moved Parser\Media -> Http\Parser
